+++ tests/no_features/from/short_nested_duplicate.rs
@@ -51,3 +51,18 @@ impl SomeTrait<member1, member2> for (){}
 )]//duplicate_end
 impl SomeTrait<member1, member2> for (){}
 //item_end
+
+#[duplicate_item(
+	name		some_int;
+	[outer_1]	[1];
+	[outer_2]	[2]
+)]//duplicate_end
+fn name()
+{
+	substitute! ( [
+		sub [	some_int	]
+	]
+		sub;
+	)
+}
+//item_end